test(web-server): share the approval store battery across backends
Factor the backend-agnostic approval store checks out of the Redis
integration test into tests/common, so any backend can run the same
conformance battery. The Redis test keeps its Docker-gated cases and
calls into the shared battery for the rest.

Ref: #271
